修复bug

+++ src/OnlineEditor.php
@@ -7,10 +7,12 @@
 
 namespace melodyne\wfm;
 
-class OnlineEditor{
+class OnlineEditor
+{
 
     //设置全局变量路径
     public $filePath = null;
+
     //设置过滤信息
     private $fileFilter = array(
         '.',
@@ -20,21 +22,24 @@ class OnlineEditor{
     );
 
     //构造函数必须是私有的在单例设计模式中
-    function __construct($filePath){
+    function __construct($filePath)
+    {
         $this->filePath = $filePath;
     }
 
     //当本类销毁的时候进行的操作
-    function __destruct(){
+    function __destruct()
+    {
         // echo $this->filePath;
     }
 
     //获取文件的内容
-    function getContent($filePath){
+    function getContent($filePath)
+    {
         if (!isset($filePath)) {
 
-        } else{
-            if(filetype($filePath)=='file'){
+        } else {
+            if (filetype($filePath) == 'file') {
                 $fileContent = file_get_contents($filePath);
                 return $fileContent;
             }
@@ -42,52 +47,74 @@ class OnlineEditor{
     }
 
     //放入文件内容
-    function putContent($filePath,$fileContent){
+    function putContent($filePath, $fileContent)
+    {
+        if(substr($filePath,0,strlen($this->filePath)) !== $this->filePath){
+            die('无权操作，非配置目录的文件');
+        }
         file_put_contents($filePath, $fileContent);
     }
 
     //判断目录是否存在
-    private function judgeExist(){
+    private function judgeExist()
+    {
         //判断目录是否为空或者没有文件
-        if(is_dir($this->filePath) && file_exists($this->filePath)){
+        if (is_dir($this->filePath) && file_exists($this->filePath)) {
             return true;
-        } else{
+        } else {
             return false;
         }
     }
 
     //创建文件
-    function createFile($filename){
-        if(!file_exists($filename)){
-            fopen($filename, "w+");
+    function createFile($filename)
+    {
+        if(substr($filename,0,strlen($this->filePath)) !== $this->filePath){
+            die('无权操作，非配置目录的文件');
         }
-
-        else{
-            echo "<a href = 'index.php' >点此返回</a>";
-            die("文件已经存在,".$filename);
+        if (!file_exists($filename)) {
+            fopen($filename, "w+");
+        } else {
+            echo "<a href=\"javascript:\" onclick=\"self.location=document.referrer;\" >点此返回</a>";
+            die("文件已经存在," . $filename);
         }
 
     }
+
     //删除文件
-    function delFile($filename){
-        if(file_exists($filename)){
-            unlink($filename);
+    function delFile($filename)
+    {
+        if(substr($filename,0,strlen($this->filePath)) !== $this->filePath){
+            die('无权操作，非配置目录的文件');
+        }
+        if (file_exists($filename)) {
+            if(is_dir($filename)){
+                if(!rmdir($filename)){
+                    echo '删除失败，可能文件下有文件';
+                    die();
+                }
+            }else{
+                unlink($filename);
+            }
+        }else{
+            die('文件不存在：'.$filename);
         }
     }
 
     // 文件排序
-    function my_sort($arrays,$sort_key,$sort_order=SORT_ASC,$sort_type=SORT_NUMERIC ){
-        if($arrays==[])return [];
-        if(!is_array($arrays))return false;
+    function my_sort($arrays, $sort_key, $sort_order = SORT_ASC, $sort_type = SORT_NUMERIC)
+    {
+        if ($arrays == []) return [];
+        if (!is_array($arrays)) return false;
 
-        foreach ($arrays as $array){
-            if(is_array($array)){
+        foreach ($arrays as $array) {
+            if (is_array($array)) {
                 $key_arrays[] = $array[$sort_key];
-            }else{
+            } else {
                 return false;
             }
         }
-        array_multisort($key_arrays,$sort_order,$sort_type,$arrays);
+        array_multisort($key_arrays, $sort_order, $sort_type, $arrays);
         return $arrays;
     }
 
@@ -108,8 +135,7 @@ class OnlineEditor{
             3 => " GB",
             4 => " TB"
         );
-        if ($size == 0)
-        {
+        if ($size == 0) {
             return str_repeat(" ", $prec) . "0$units[0]";
         }
         $unit = min(4, floor(log($size) / log(2) / 10));
@@ -121,8 +147,9 @@ class OnlineEditor{
     }
 
     //主函数
-    function main(){
-        if($this->judgeExist()){
+    function main()
+    {
+        if ($this->judgeExist()) {
             //获取打开文件夹对象
             $fileOpen = opendir($this->filePath);
             $dirArr = array();
@@ -131,37 +158,37 @@ class OnlineEditor{
             //遍历文件夹
             while ($file = readdir($fileOpen)) {
                 //过滤
-                if(in_array($file, $this->fileFilter)){
+                if (in_array($file, $this->fileFilter)) {
                     continue;
                 }
-                $path = rtrim($this->filePath,'/').'/'.$file;
+                $path = rtrim($this->filePath, '/') . '/' . $file;
                 $type = fileType($path);
                 $size = $this->formatSize(fileSize($path));
-                if($type=='dir'){ // 文件夹
+                if ($type == 'dir') { // 文件夹
                     $dirArr[] = array(
-                        'fileCode'  => $i,
-                        'fileName'  => $file,
-                        'fileType'  => $type,
-                        'fileSize'  => $size,
+                        'fileCode' => $i,
+                        'fileName' => $file,
+                        'fileType' => $type,
+                        'fileSize' => $size,
                         'filemtime' => filemtime($path)
                     );
-                }else{ // 文件
+                } else { // 文件
                     $fileArr[] = array(
-                        'fileCode'  => $i,
-                        'fileName'  => $file,
-                        'fileType'  => $type,
-                        'fileSize'  => $size,
+                        'fileCode' => $i,
+                        'fileName' => $file,
+                        'fileType' => $type,
+                        'fileSize' => $size,
                         'filemtime' => filemtime($path)
                     );
                 }
                 $i++;
             }
-            $dirArr = $this->my_sort($dirArr,'fileName',SORT_ASC,SORT_STRING);
-            $fileArr = $this->my_sort($fileArr,'fileName',SORT_ASC,SORT_STRING);
+            $dirArr = $this->my_sort($dirArr, 'fileName', SORT_ASC, SORT_STRING);
+            $fileArr = $this->my_sort($fileArr, 'fileName', SORT_ASC, SORT_STRING);
             closedir($fileOpen); //关闭文件
-            return array_merge($dirArr,$fileArr);
+            return array_merge($dirArr, $fileArr);
 
-        } else{
+        } else {
             die("不存在此文件夹");
         }
     }
